feat: tokenUtils 파일 삭제
